requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==25.1.0; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
coreapi==2.3.3
coreschema==0.0.4
django==5.1.6; python_version >= '3.10'
django-rest-swagger==2.2.0
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.4.0; python_version >= '3.9'
drf-spectacular==0.28.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.5; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
markupsafe==3.0.2; python_version >= '3.9'
openapi-codec==1.3.2
pyjwt==2.10.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
rpds-py==0.22.3; python_version >= '3.9'
simplejson==3.19.3; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.13'
tzdata==2025.1; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.9'
